requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-admin-thumbnails==0.2.8
django-ckeditor==6.7.0
django-import-export==3.3.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-mptt==0.15.0
django-template-maths==0.2.0
et-xmlfile==1.1.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.6
magic-filter==1.0.12
MarkupPy==1.14
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
